#DSC-350
The feature_preparation.py file contains the code I used to transform the categorical data in the filtered dataset to numerical columns. It also contains the code I used to fill the NAs in the dataset.

The load_data.py file has been modified to split the original MTN metadata into 3 parts instead of just 2.